<rdar://problem/31192097> Improve caching of auth role data (#820)
To avoid having to perform lookups via the IRoleManager for every
isSuperuser or canLogin query, make RolesCache a repository of actual
Role objects, rather than RoleResources. Also adds a new method to
IRoleManager, with a default impl for compatibility, to retrieve the set
of granted roles for a primary role, represented as these new Role
objects. The default impl is pretty inefficient, essentially performing
the same multiple IRM lookups per granted role as the original version.