second round of feedback from BES
takePermit is used by shutdownNow to clear the work queue, if
resizing taking place while that happens it will stop early.
Break out the check for whether the SEPExecutor is resizing or stopping
into a separate function.

Make a couple of the SEPExecutor members private as they no longer
need to be public for metrics.