Defer initialization of stage executors.
Using Verb to filter in-JVM DTests initializes the Verb enum,
which initializes the Stage enum which tries to create thread pools
on the main in-jvm runner thread (plus DatabaseDescriptor shouldn't
be initialized so it crashes with an NPE).

Instead, initialize the executors in the stage lazily so it
is safe to use without side effects until the executor is retrieved.