rdar://59309912 (Repaired data tracking should attempt to read the same amount of data from each replica) (#1485) (#1491)
Handle cases of both under and overreading of the repaired set
due to variances in the amount of unrepaired data on a replica.
A replica will attempt to read a fixed amount of repaired data,
based on the query's DataLimit, while generating the repaired
digest. If less than that amount is read in the course of satisfying
the query limits, we overread. If more that the amount is read
during the main read phase (because of unrepaired deletions
shadowing repaired data), the digest calculation stops when its
own limit is reached.