rdar://67760856 (CIE4 Out of token range patch does not handle PartitionRangeReadCommands) (#1878)
Adds a new `onePartitionKey()` method to ReadCommand and refactors the out of range token
patch to use that.  Previously in 3.0 the OOTR patch just checked the ReadCommand.Kind
was a single partition, but in the 4.0 port it was updated to use the newer `isLimitedToOnePartition` which can be true for `PartitionRangeReadCommand`s to.  

For now, revert to the more conservative 3.0 version. May review a more comprehensive fix once `teslaperf` cluster is operational again.