Backport CASSANDRA-14297: rdar://62201329 Read Unavailables during bounces of SSL Enabled clusters (#2142)
* rdar://62201329 Read Unavailables during bounces of SSL Enabled clusters

This improves on the wait for healthy work from CASSANDRA-13993 to
solve CASSANDRA-14297. In particular now the connectivity checker waits
for all but a single node in either the local datacenter or every
datacenter (defaults to just local, but the user can configure it to
wait for every datacenter). This way users can use this feature to ensure
availability of their application during restarts of Cassandra. The default
behavior waits for all but a single local datacenter node.

Patch by Joseph Lynch; Reviewed by Ariel Weisberg for CASSANDRA-14297

* added warning for voided property

* fix deprecation of block_for_peers_percentage

* correct comment

Co-authored-by: Joseph Lynch <joe.e.lynch@gmail.com>
Co-authored-by: Marcus Eriksson <marcus_eriksson@apple.com>