Forward port CASSANDRA-16422 Filter decomissioned/assassinated nodes when checking if safe to alter keyspace
As part of Transient Replication added in CASSANDRA-14404, ALTER KEYSPACE
had a validation check added to verify no range movement was taking place.
The check requires all nodes to be in NORMAL state. When a node is
decommissioned or assassinated then it will remain in gossip for 3 days in
LEFT state (by default), preventing any keyspace alteration until it has expired from gossip, even though it is safe to make the change.

This patch filters the unreachable nodes to remove any LEFT nodes.

Patch by Jon Meredith; reviewed by ? for CASSANDRA-16422

Should be removed in the next OSS rebase.