rdar://75617400 (Prevent PrepareMessage#serializedSize() from throwing an assertion in mixed-mode)
If serializedSize() throws, onFailedSerialize() also does, causing some of necessary OutboundConnection#doRun() logic to not be executed.

Throwing in serializedSize() is unexpected and should never happen. Indeed, it only does so in PrepareMessage serialised, when explicitly asserting a version (this is holdover logic from pre-4.0 days, that was carried over blindly).

Removing the assertion from serializedSize() implementation of PrepareMessage serialised will prevent OutboundConnection logic from choking in mixed mode, when attempting to send a PrepareMessage from a 4.0 node to 3.0 one.