(Early Port, updated) Avoid memoizing the wrong min cluster version during upg… (#2479) (#2482)
The early port of CASSANDRA-16759 was done a little too early and the upstream version
ended up having some changes made on commit to pass a failing test.  Rio did not fail
our internal build when building the release so the failure went undetected.

This patch gets as close as possible to the final upstream version
(2fba5c80ce7bf71d04c62043ffa1088b9e832d83) by pulling in a few other Gossiper fixes
of the next rebase.

Both this patch and the first version of it should drop out next time they are rebased.