Fix downstream pipeline triggering (#2578)
See recent failure: https://rio.apple.com/projects/cie-cassandra/pipeline-specs/cie-cassandra-if1-release/pipelines/43962984-d70d-49bf-8824-b110c52f288f/log

This run failed due to a missing file which intentionally does not exist
on the 3.0.24.x branch but does on 4.0.0.x. The default Rio behavior
passes the git SHA from an upstream pipeline to a downstream one,
causing the pipeline to use a branch missing the necessary file.

This is not an issue for the 4.0.0.x branch.