rdar://84756120 (CRITICAL : iccyruscassandra - Read Timeouts) (#2645)
If the repair coordinator gets a FailSession message after it has marked the
session as FINALIZED it will start this infinite loop of FailSession messages

This patch avoids getting in to the infinite loop of messages, but we still don't know
exactly how this happened in iccyrus. This will help us debug as logs won't get rotated
out in seconds.

(rebased-with-history from commit 91075de32012628d84b49a78f12c734f24b0a149)