rdar://84756120 (CRITICAL : iccyruscassandra - Read Timeouts) (#2645)
If the repair coordinator gets a FailSession message after it has marked the
session as FINALIZED it will start this infinite loop of FailSession messages

This patch avoids getting in to the infinite loop of messages, but we still don't know
exactly how this happened in iccyrus. This will help us debug as logs won't get rotated
out in seconds.