rdar://84106393 (4.0: Investigate Xmas patch last repair time being updated with the current value error)
Downgrade the error to an info message if the same time.  Ranges were being
updated with the same time when handling APPLE_REPAIRED_RANGES_REQ which should have the same
time if the two hosts are up to date.

Now if the times are the same it issues a DEBUG level message which does not appear in the exception
reports or trip python dtests expecting no ERRORs.
```
TRACE [Stream-Deserializer-/127.0.0.1:7000-22eee005] 2021-12-08T21:25:51,625 MessagingService.java:321 - /127.0.0.4:7000 sending APPLE_REPAIRED_RANGES_REQ to 161@/127.0.0.1:7000
DEBUG [MiscStage:1] 2021-12-08T21:25:51,628 ColumnFamilyStore.java:1863 - Kept same last repair time for range (00030,00010] from 1638998645 to 1638998645
```