fixup! rdar://84012150 (Mixed cluster rewrote sstable metadata with older format, reporting corruption on upgrade) #2597
Disabling this test if the upgrade-from writes current sstable format files.
Not sure how valuable this test will be going forward with how frequently dtest-api changes
mean it is not possible to run against older dtest jars.

The test has to wait for the cluster to be created before it knows the version, so
spends a minute or so to do nothing of value if the test should be skipped.