In-JVM internode messing fix - MixedModeFrom3LoggedBatchTest.testSimpleStrategy
Test is failing because of a message being delivered before token metadata
is fully initialized and giving an exception about the ring being empty.

When mock messaging is being used, messages can be delivered before the
messaging service would be initialized.  Add some protection to only
enable delivery by mock messaging once it is set up and the filters
are created.