rdar://95731143 (BulkLoader initializes system keyspaces)
After CASSANDRA-17663 using the bulk loader will initialize the system keyspace to try and retrieve the preferred IP.

Upstream may be a little different as there’s been some movement to deprecate the old ssl storage port flag.
This patch should empty out if upstream is fixed and this branch is rebased.