rdar://100526479 (StorageService.getNativeaddress does not account for IPv6 addresses in the case NATIVE_ADDRESS_AND_PORT is not present in gossip state for an endpoint)
Fix StorageService.getNativeaddress handling of IPv6 addresses
StorageService.getNativeaddress does not currently correctly handle
IPv6 addresses correctly when NATIVE_ADDRESS_AND_PORT are not present in
that it simply concatenates the IP address with the default native port,
e.g.:

0:0:0:0:0:0:5a:3:9042

This does not parse into an InetSocketAddress as the address and port
can't be disambiguated.

Such a case would usually be present when there are 3.x nodes present in a
cluster with 4.0 nodes.

Change updates RPC_ADDRESS and else case to create InetAddressAndPort instances
with DatabaseDescriptor.getNativeTransportPort and returns the
getHostAddress(withPort) which properly bracket encodes the address,
e.g.:

[0:0:0:0:0:0:5a:3]:9042

which can be parsed as an InetSocketAddress.

patch by Andy Tolbert; reviewed by Ariel Weisberg, Brandon Williams for CASSANDRA-17945